Fix "pure" settings in meson.build
python-rtmidi installed the extension and pure-Python sources in the same
directory without setting "pure: false" anywhere in meson.build. Since
version 0.13.0, such behaviour causes meson-python to throw an error.
Fixed by setting "pure: false" for the whole project (which necessitated
bumping the minimum Meson version from 0.63.0 to 0.64.0) and removing
the one occurrence of "pure: true" (which was redundant to begin with
- true is the default).

Signed-off-by: Marek Szuba <marecki@gentoo.org>